First implementation of arguments.
Goliath server will check for the existance of at least one argument. It
takes the address of the Schneider and Maxon motors. If it has at least
one of those it will try to initialize the motors. Otherwise, it acts
like a TCPIP server for troubleshooting, and does not touch any serial
device or motor.

Note that the Maxon code is still incomplete!